+++ scripts/tagger.py
@@ -3,7 +3,6 @@ from PIL import Image, ImageFile
 from modules import script_callbacks
 from tagger.api import on_app_started
 from tagger.ui import on_ui_tabs
-from tagger.settings import on_ui_settings
 
 
 # if you do not initialize the Image object
@@ -17,4 +16,3 @@ ImageFile.LOAD_TRUNCATED_IMAGES = True
 
 script_callbacks.on_app_started(on_app_started)
 script_callbacks.on_ui_tabs(on_ui_tabs)
-script_callbacks.on_ui_settings(on_ui_settings)
