+++ scripts/tagger/settings.py
